implement weather forecast and details

+++ app/src/main/java/com/kasko/victor/weather_places/domain/model/Temperature.kt
@@ -1,6 +1,6 @@
 package com.kasko.victor.weather_places.domain.model
 
 data class Temperature(
-    val day: Int,
-    val night: Int
+    val day: Double,
+    val night: Double
 )
